beginner/weather-api/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
coverage==7.13.5
Deprecated==1.3.1
fastapi==0.136.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
iniconfig==2.3.0
librt==0.9.0
limits==5.8.0
mypy==1.20.2
mypy_extensions==1.1.0
packaging==26.2
pathspec==1.1.1
pluggy==1.6.0
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
pytest==9.0.3
pytest-cov==7.1.0
pytest-mock==3.15.1
python-dotenv==1.2.2
redis==7.4.0
requests==2.33.1
slowapi==0.1.9
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.46.0
wrapt==2.1.2
